refactor: 바텀 모달을 불러오는 코드가 여러 코드에서 중복 사용되어 utils 폴더로 분리 #55
